# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inkweld API",
    "version": "1.0.0",
    "description": "Collaborative creative writing platform API (Bun)"
  },
  "servers": [
    {
      "url": "http://localhost:8333",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "An error occurred"
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique project identifier"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly project identifier"
          },
          "title": {
            "type": "string",
            "description": "Project title"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Project description"
          },
          "username": {
            "type": "string",
            "description": "Project owner username"
          },
          "coverImage": {
            "type": "string",
            "nullable": true,
            "description": "Cover image URL"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "description": "Project creation date"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "username",
          "createdDate",
          "updatedDate"
        ]
      },
      "CreateProjectRequest": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 3,
            "pattern": "^[a-z0-9-]+$",
            "description": "URL-friendly project identifier (lowercase, numbers, hyphens only)",
            "example": "my-novel"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "description": "Project title",
            "example": "My Novel"
          },
          "description": {
            "type": "string",
            "description": "Optional project description"
          }
        },
        "required": [
          "slug",
          "title"
        ]
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "doc-123",
            "description": "Document ID"
          },
          "name": {
            "type": "string",
            "example": "Chapter 1",
            "description": "Document name"
          },
          "type": {
            "type": "string",
            "example": "ITEM",
            "description": "Document type"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "type"
        ]
      },
      "DocumentError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "Element": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "elem-123",
            "description": "Element ID"
          },
          "name": {
            "type": "string",
            "example": "Chapter 1",
            "description": "Element name"
          },
          "type": {
            "type": "string",
            "enum": [
              "FOLDER",
              "ITEM",
              "CHARACTER",
              "LOCATION",
              "WB_ITEM",
              "MAP",
              "RELATIONSHIP",
              "PHILOSOPHY",
              "CULTURE",
              "SPECIES",
              "SYSTEMS"
            ],
            "example": "ITEM",
            "description": "Element type"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "Parent element ID"
          },
          "order": {
            "type": "number",
            "example": 0,
            "description": "Order in parent"
          },
          "level": {
            "type": "number",
            "example": 0,
            "description": "Nesting level in tree hierarchy"
          },
          "expandable": {
            "type": "boolean",
            "example": false,
            "description": "Whether element can be expanded (folders)"
          },
          "version": {
            "type": "number",
            "example": 1,
            "description": "Version number for optimistic locking"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Element metadata key-value pairs"
          },
          "createdAt": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "parentId",
          "order",
          "level",
          "expandable",
          "version",
          "metadata"
        ]
      },
      "ElementError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "ProjectFile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "document.pdf",
            "description": "File name"
          },
          "size": {
            "type": "number",
            "example": 1024,
            "description": "File size in bytes"
          },
          "uploadDate": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Upload timestamp"
          }
        },
        "required": [
          "name"
        ]
      },
      "FileError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "UploadResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "document.pdf",
            "description": "Uploaded file name"
          },
          "size": {
            "type": "number",
            "example": 1024,
            "description": "File size in bytes"
          },
          "uploadDate": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Upload timestamp"
          }
        },
        "required": [
          "name",
          "size",
          "uploadDate"
        ]
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "File deleted successfully",
            "description": "Success message"
          }
        },
        "required": [
          "message"
        ]
      },
      "EpubResponse": {
        "type": "object",
        "properties": {
          "originalName": {
            "type": "string",
            "example": "project.epub",
            "description": "Original filename"
          },
          "storedName": {
            "type": "string",
            "example": "project-123.epub",
            "description": "Stored filename"
          },
          "contentType": {
            "type": "string",
            "example": "application/epub+zip",
            "description": "MIME type"
          },
          "size": {
            "type": "number",
            "example": 102400,
            "description": "File size in bytes"
          },
          "uploadDate": {
            "type": "string",
            "example": "2023-01-01T00:00:00.000Z",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "originalName",
          "storedName",
          "contentType",
          "size",
          "uploadDate"
        ]
      },
      "EpubError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "listUserProjects",
        "responses": {
          "200": {
            "description": "List of user projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or project already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/:username/:slug/docs": {
      "get": {
        "operationId": "listProjectDocuments",
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentError"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/:username/:slug/docs/:docId": {
      "get": {
        "operationId": "getProjectDocument",
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "doc-123",
              "description": "Document ID"
            },
            "required": true,
            "description": "Document ID",
            "name": "docId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Document metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentError"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/:username/:slug/elements": {
      "get": {
        "operationId": "listProjectElements",
        "tags": [
          "Elements"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List of project elements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Element"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElementError"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElementError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/:username/:slug/files": {
      "get": {
        "operationId": "listProjectFiles",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "List of files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectFile"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "uploadProjectFile",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "No file provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/:username/:slug/files/:storedName": {
      "get": {
        "operationId": "downloadProjectFile",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "document.pdf",
              "description": "File name"
            },
            "required": true,
            "description": "File name",
            "name": "storedName",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteProjectFile",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "document.pdf",
              "description": "File name"
            },
            "required": true,
            "description": "File name",
            "name": "storedName",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          },
          "404": {
            "description": "File or project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/:username/:slug/epub": {
      "post": {
        "tags": [
          "Export"
        ],
        "operationId": "exportProjectAsEpub",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "johndoe",
              "description": "Username"
            },
            "required": true,
            "description": "Username",
            "name": "username",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "example": "my-novel",
              "description": "Project slug"
            },
            "required": true,
            "description": "Project slug",
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "EPUB export metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubError"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubError"
                }
              }
            }
          },
          "500": {
            "description": "Export failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubError"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpubError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}